% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badpkg.R
\name{badpkg}
\alias{badpkg}
\alias{import_as2}
\alias{import_inops2}
\alias{import_int2}
\alias{import_LL2}
\title{badpkg}
\usage{
import_as2(...)

import_inops2(...)

import_int2(...)

import_LL2(...)
}
\arguments{
\item{A}{whatever}

\item{X}{whatever}
}
\description{
Bla
}
